Вносит шаблоны в компоненты
